Added shcodecs-record tool
This tool encodes v4l2 camera data and shows a preview on the display. It allows the camera images to be rotated & cropped before encoding for use with boards where the camera orientation is different to the display, e.g. MigoR.

The data flow could be improved if the camera captured straight to the VPU encoder input buffers. If implemented, rotation of the camera image would not be supported.