Fixed thread sync issues with shcodecs-play
Noticed that the playback sometimes appeared to show frames out
of order. This was visible on the Algo and ms7724se boards whilst
playing a QVGA stream.

The way the scale thread was synchronised to the decoder output
meant that either mutex could be locked multiple times by a thread
before being unlocked by the other thread, thus breaking the
intended synchronisation.

This fix uses posix conditional variables for synchronisation.