rename sdr_{read,write} src,dest args
It is not clear from the types which arguments are physical or virtual,
and the argument order of read differs from convention (eg. memcpy).
For each function, clarify which argument is src and which is dest, and
tag each as physical or virtual.

Backported from b3a074c57f831c1d8837dcaea7393208ae84ae5d (uiomux branch)