Revert "defer VPU temporary buffer allocation"
This reverts commit e91af41836ca4dcb678cc11e04c55b489096353a.

Problem: The temporary buffer allocation was being deferred until after the
vpu4 params were set, so it was ignored by the middleware.
Symptoms: The encoder appeared to run fine, but only a solid grey image was
produced.
Issues: Reverting this patch breaks the ability to set width,height
after initialization. All of the initialization should be deferred in
order to allow this functionality. In any case there is not yet a
separate function for intitializing the encoder without setting width
and height, so this revert does not break any existing applications.