Fixed crap output data and infinity loop in decoder
Two issues, one was that the input buffer had to be a multiple of 32 bytes.
It doesn't appear to need to be zero padded. The other was that, prior to
fixing the buffer length alignment, the decoder returned an error but the
decoder kept chucking the same data at the vpu.